+++ BaseFramework/src/com/base/web/organization/OrganizationDao.java
@@ -2,7 +2,7 @@ package com.base.web.organization;
 
 import com.base.core.ssh.l3dao.BaseDao;
 
-public interface OrganizationDao extends BaseDao<Organization,Integer>
+public interface OrganizationDao extends BaseDao<Organization,Long>
 {
 
 }
