+++ BaseFramework/src/com/base/web/organization/OrganizationDaoImp.java
@@ -2,7 +2,7 @@ package com.base.web.organization;
 
 import com.base.core.ssh.l3dao.BaseDaoImp;
 
-public class OrganizationDaoImp extends BaseDaoImp<Organization,Integer> implements OrganizationDao
+public class OrganizationDaoImp extends BaseDaoImp<Organization,Long> implements OrganizationDao
 {
 
 	public OrganizationDaoImp()
