+++ BaseFramework/src/com/base/web/permission/FunctionDao.java
@@ -2,7 +2,7 @@ package com.base.web.permission;
 
 import com.base.core.ssh.l3dao.BaseDao;
 
-public interface FunctionDao extends BaseDao<Function,Integer>
+public interface FunctionDao extends BaseDao<Function,Long>
 {
 
 }
