+++ BaseFramework/src/com/base/web/permission/FunctionDaoImp.java
@@ -2,7 +2,7 @@ package com.base.web.permission;
 
 import com.base.core.ssh.l3dao.BaseDaoImp;
 
-public class FunctionDaoImp extends BaseDaoImp<Function,Integer> implements FunctionDao
+public class FunctionDaoImp extends BaseDaoImp<Function,Long> implements FunctionDao
 {
 
 	public FunctionDaoImp()
