+++ BaseFramework/src/com/base/web/permission/PermissionDao.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import com.base.core.ssh.l3dao.BaseDao;
 
-public interface PermissionDao extends BaseDao<Permission,Integer>
+public interface PermissionDao extends BaseDao<Permission,Long>
 {
 
 	List<Function> queryFunction(String hql);
