+++ BaseFramework/src/com/base/web/permission/ResourceDao.java
@@ -2,7 +2,7 @@ package com.base.web.permission;
 
 import com.base.core.ssh.l3dao.BaseDao;
 
-public interface ResourceDao extends BaseDao<Resource,Integer>
+public interface ResourceDao extends BaseDao<Resource,Long>
 {
 
 }
