+++ BaseFramework/src/com/base/web/permission/ResourceDaoImp.java
@@ -2,13 +2,12 @@ package com.base.web.permission;
 
 import com.base.core.ssh.l3dao.BaseDaoImp;
 
-public class ResourceDaoImp extends BaseDaoImp<Resource,Integer> implements ResourceDao
+public class ResourceDaoImp extends BaseDaoImp<Resource,Long> implements ResourceDao
 {
 
 	public ResourceDaoImp()
 	{
 		super(Resource.class);
 	}
-
 
 }
