+++ BaseFramework/src/com/base/web/permission/RoleDao.java
@@ -2,7 +2,7 @@ package com.base.web.permission;
 
 import com.base.core.ssh.l3dao.BaseDao;
 
-public interface RoleDao extends BaseDao<Role,Integer>
+public interface RoleDao extends BaseDao<Role,Long>
 {
 
 }
