+++ BaseFramework/src/com/base/web/permission/RoleDaoImp.java
@@ -2,7 +2,7 @@ package com.base.web.permission;
 
 import com.base.core.ssh.l3dao.BaseDaoImp;
 
-public class RoleDaoImp extends BaseDaoImp<Role,Integer> implements RoleDao
+public class RoleDaoImp extends BaseDaoImp<Role,Long> implements RoleDao
 {
 
 	public RoleDaoImp()
