+++ BaseFramework/src/com/base/web/permission/service/FunctionService.java
@@ -3,7 +3,7 @@ package com.base.web.permission.service;
 import com.base.core.ssh.l2service.BaseService;
 import com.base.web.permission.Function;
 
-public interface FunctionService extends BaseService<Function,Integer>
+public interface FunctionService extends BaseService<Function,Long>
 {
 
 }
