+++ BaseFramework/src/com/base/web/permission/service/ResourceService.java
@@ -3,7 +3,7 @@ package com.base.web.permission.service;
 import com.base.core.ssh.l2service.BaseService;
 import com.base.web.permission.Resource;
 
-public interface ResourceService extends BaseService<Resource,Integer>
+public interface ResourceService extends BaseService<Resource,Long>
 {
 
 }
