+++ BaseFramework/src_lock/com/lock/unlockInfo/UnlockInfoDao.java
@@ -2,7 +2,7 @@ package com.lock.unlockInfo;
 
 import com.base.core.ssh.l3dao.BaseDao;
 
-public interface UnlockInfoDao extends BaseDao<UnlockInfo,Integer>
+public interface UnlockInfoDao extends BaseDao<UnlockInfo,Long>
 {
 
 }
